Initial Frontend (#4)
- Generated boilerplate with `npm create vite@latest`, then pared down
everything I didn't want, need, or understand.
- Created a generic "polling state machine" hook, `usePoller`.
- Threw together a front end (all in `App.tsx`), including a "game list"
page, and a "game" page, the latter of which will poll for an updated
game state if the current player is waiting for another player to make a
choice, else allow the current player to select "options" and send the
resulting "decision" to the server.